pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "3dcolorconverter"
version = "0.1.0"
description = "Convert textured 3D models into printable color-region assets."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Phil Tullai" }
]
dependencies = [
  "numpy>=1.24",
  "Pillow>=10.0",
  "trimesh>=4.0",
  "scikit-learn>=1.3",
  "scipy>=1.10",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
3dcolorconverter = "color3dconverter.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
